move esm/ to dist/
