+++ .changeset/gold-nails-share.md
@@ -2,4 +2,4 @@
 'monaco-graphql': minor
 ---
 
-avoid bundling unnecessary languages — import `monaco-graphql/esm/monaco-editor` instead of `monaco-editor` to reduce your bundle size, as that imports only `graphql` and `json` languages and leaves out `ts`, `css`, `html`, and much more
+avoid bundling unnecessary languages — import `monaco-graphql/dist/monaco-editor` instead of `monaco-editor` to reduce your bundle size, as that imports only `graphql` and `json` languages and leaves out `ts`, `css`, `html`, and much more
