+++ examples/monaco-graphql-nextjs/src/constants.ts
@@ -1,5 +1,5 @@
-import { editor, Uri } from 'monaco-graphql/esm/monaco-editor';
-import { initializeMode } from 'monaco-graphql/esm/initializeMode';
+import { editor, Uri } from 'monaco-graphql/dist/monaco-editor';
+import { initializeMode } from 'monaco-graphql/dist/initializeMode';
 
 type ModelType = 'operations' | 'variables' | 'response';
