+++ examples/monaco-graphql-nextjs/src/editor.tsx
@@ -5,7 +5,7 @@ import {
   KeyMod,
   KeyCode,
   languages,
-} from 'monaco-graphql/esm/monaco-editor';
+} from 'monaco-graphql/dist/monaco-editor';
 import { createGraphiQLFetcher } from '@graphiql/toolkit';
 import * as JSONC from 'jsonc-parser';
 import {
