+++ examples/monaco-graphql-react-vite/vite.config.ts
@@ -11,7 +11,7 @@ export default defineConfig({
       customWorkers: [
         {
           label: 'graphql',
-          entry: 'monaco-graphql/esm/graphql.worker',
+          entry: 'monaco-graphql/dist/graphql.worker',
         },
       ],
     }),
