+++ examples/monaco-graphql-webpack/src/editors.ts
@@ -1,4 +1,4 @@
-import { editor, Uri } from 'monaco-graphql/esm/monaco-editor';
+import { editor, Uri } from 'monaco-graphql/dist/monaco-editor';
 
 const GRAPHQL_LANGUAGE_ID = 'graphql';
