+++ examples/monaco-graphql-webpack/src/index.ts
@@ -1,8 +1,8 @@
 /* global netlify */
 
-import { editor, KeyMod, KeyCode } from 'monaco-graphql/esm/monaco-editor';
+import { editor, KeyMod, KeyCode } from 'monaco-graphql/dist/monaco-editor';
 import * as JSONC from 'jsonc-parser';
-import { initializeMode } from 'monaco-graphql/esm/initializeMode';
+import { initializeMode } from 'monaco-graphql/dist/initializeMode';
 
 import { createEditors } from './editors';
 import { schemaFetcher, schemaOptions } from './schema';
