+++ packages/graphiql-toolkit/package.json
@@ -15,8 +15,8 @@
     "url": "https://github.com/graphql/graphiql/issues?q=issue+label:@graphiql/toolkit"
   },
   "license": "MIT",
-  "module": "esm/index.js",
-  "types": "esm/index.d.ts",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {},
   "dependencies": {
     "@n1ru4l/push-pull-async-iterable-iterator": "^3.1.0",
