+++ packages/graphql-language-service-cli/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../../resources/tsconfig.base.json",
   "compilerOptions": {
     "rootDir": "./src",
-    "outDir": "./esm",
+    "outDir": "./dist",
     "composite": true,
     "target": "ES2018"
   },
