+++ packages/graphql-language-service-server/package.json
@@ -30,7 +30,7 @@
     "svelte",
     "typescript"
   ],
-  "module": "esm/index.js",
+  "module": "dist/index.js",
   "typings": "esm/index.d.ts",
   "peerDependencies": {
     "graphql": "^15.5.0 || ^16.0.0"
