+++ packages/graphql-language-service/package.json
@@ -25,8 +25,8 @@
     "language service",
     "LSP"
   ],
-  "module": "esm/index.js",
-  "types": "esm/index.d.ts",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "bin": {
     "graphql": "./dist/temp-bin.js"
   },
